# api/openapi.yaml
openapi: 3.1.0
info:
  title: Distributed HF Downloader API
  version: 2.0.0
  description: |
    Distributed HuggingFace model weight downloader.

    Authentication:
      - User APIs: OIDC + Bearer JWT
      - Executor APIs: mTLS + Bearer JWT + HMAC body signature
      - Admin APIs: same as User APIs but require role=tenant_admin or system_admin

    Tenancy:
      - All resources are tenant-scoped except system-level admin endpoints
      - Header `X-Tenant-Slug` only honored for system_admin role

    Rate limits:
      - Per-route token bucket (see 02-protocol.md §8.1)

    Error format:
      - All errors: `{code, message, trace_id, details?}`

  contact:
    email: dlw@example.com
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://api.dlw.example.com/api/v2
    description: Production
  - url: https://staging-api.dlw.example.com/api/v2
    description: Staging
  - url: http://localhost:8000/api/v2
    description: Local dev
...
security:
  - bearerAuth: []
paths:
...
  /executors/register:
    post:
      tags: [executors]
      summary: Register an executor (mTLS bootstrap)
      description: |
        Called once during executor startup. Requires enrollment_token (out-of-band).
        Returns mTLS client certificate and JWT.
      operationId: registerExecutor
      security: []
      parameters:
        - in: header
          name: X-Enrollment-Token
          required: true
          schema: {type: string}
          description: 256-bit hex enrollment secret
        - in: header
          name: X-Enrollment-Nonce
          required: true
          schema: {type: string}
          description: 128-bit random hex
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ExecutorRegisterRequest'}
      responses:
        '200':
          description: Registered
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExecutorRegisterResponse'}
        '401':
          description: Invalid enrollment token
        '409':
          description: Nonce replay or executor_id conflict
...
  /executors/{executorId}/renew:
    parameters:
      - in: path
        name: executorId
        required: true
        schema: {type: string}
    post:
      tags: [executors]
      summary: Renew executor JWT and optionally rotate mTLS cert
      operationId: renewExecutorJwt
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                client_csr_pem:
                  type: string
                  nullable: true
                  description: X.509 CSR PEM for cert rotation; omit or null to renew JWT only
      responses:
        '200':
          description: Renewed credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  executor_jwt: {type: string}
                  jwt_renew_in_seconds: {type: integer}
                  client_cert_pem:
                    type: string
                    nullable: true
                    description: New PEM-encoded client cert; null if no CSR was provided
                  cert_renew_in_seconds:
                    type: integer
                    nullable: true
                    description: Seconds until next cert renewal; null if cert not rotated
        '401':
          description: Invalid or expired executor JWT / mTLS cert
  /executors/{executorId}/poll:
    parameters:
      - in: path
        name: executorId
        required: true
        schema: {type: string}
      - in: header
        name: X-Executor-Epoch
        required: true
        schema:
          type: integer
          minimum: 1
        description: Fence token. Must match the executor's current epoch.
    post:
      tags: [executors]
      summary: Poll for next subtask assignment
      operationId: pollForSubtask
      responses:
        '200':
          description: Assignment (or empty if no work available)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AssignmentResponse'}
        '401':
          description: Epoch mismatch (stale executor) or bearer token invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    oneOf:
                      - type: string
                      - type: object
                        properties:
                          code:
                            type: string
                            enum: [EPOCH_MISMATCH]
                          expected:
                            type: integer
                          got:
                            type: integer
        '503':
          description: Controller is recovering after a failover (INVARIANT 33). Retry shortly.
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: object
                    properties:
                      code: { type: string, enum: [CONTROLLER_RECOVERING] }
                      message: { type: string }
...
components:
  securitySchemes:
    # CODE-09 修复 (v2.0.16): 完整表达 modelpull 的 3 套鉴权
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User-side OIDC-issued JWT (UI / CLI / SDK)
    mtlsAuth:
      type: mutualTLS
      description: |
        OAS 3.1 native mTLS scheme. Used for executor → controller channel
        in addition to bearerAuth (executor JWT). Server validates client
        cert fingerprint against `executors.cert_fingerprint`.
    hmacBodySig:
      type: apiKey
      in: header
      name: X-Heartbeat-HMAC
      description: |
        HMAC-SHA256 of (request body + nonce + timestamp) using executor's
        enrollment_secret. Required for /executors/heartbeat in addition to
        mTLS + bearerAuth (defense in depth). Detail see 04 §2.2.4.
...
  schemas:
...
    ExecutorRegisterRequest:
      type: object
      required: [host_id, executor_id_proposal, capabilities, client_csr_pem]
      properties:
        host_id: {type: string}
        executor_id_proposal: {type: string}
        tenant_id: {type: integer, format: int64, nullable: true}
        parent_executor_id: {type: string, nullable: true}
        display_name: {type: string, maxLength: 128, nullable: true}
        location: {type: string, maxLength: 128, nullable: true}
        tags:
          type: array
          items: {type: string}
        capabilities:
          type: object
          properties:
            nic_speed_gbps: {type: integer}
            regions:
              type: array
              items: {type: string}
            storage_backends:
              type: array
              items: {type: integer}
        credential_pool_summary:
          type: object
          description: |
            Aliases only; secrets stay on executor (invariant 30, v2.1).
            See 14 §3.3 for full protocol.
          properties:
            gateway_aliases:
              type: array
              items: {type: string}
            hf_token_aliases:
              type: array
              items: {type: string}
            s3_credential_aliases:
              type: array
              items:
                type: object
                properties:
                  alias: {type: string}
                  source_id: {type: string}
        client_csr_pem:
          type: string
          description: X.509 CSR in PEM format
    ExecutorRegisterResponse:
      type: object
      properties:
        executor_id: {type: string}
        epoch: {type: integer, format: int64}
        client_cert_pem: {type: string, description: PEM-encoded X.509 client certificate}
        ca_chain:
          type: array
          items: {type: string}
          description: CA certificate chain PEM strings
        executor_jwt: {type: string}
        hmac_seed_hex: {type: string, description: 256-bit hex HMAC seed for heartbeat signatures}
        cert_renew_in_seconds: {type: integer, description: Seconds until certificate renewal recommended}
        jwt_renew_in_seconds: {type: integer, description: Seconds until JWT renewal recommended}
        jwt_signing_alg: {type: string, const: EdDSA}
        next_renew_in_seconds: {type: integer, description: "Deprecated: use cert_renew_in_seconds"}
...
    StorageConfig:
      type: object
      required: [bucket]
      properties:
        bucket:
          type: string
        region:
          type: string
          default: us-east-1
        endpoint_url:
          type: string
          nullable: true
        key_prefix:
          type: string
          default: ""
    SubTaskRead:
      type: object
      required: [id, task_id, filename, status]
      description: Subtask detail returned in assignment payload and GET subtask detail.
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        filename:
          type: string
          maxLength: 512
        file_size:
          type: integer
          format: int64
          nullable: true
        expected_sha256:
          type: string
          nullable: true
        status:
          type: string
        s3_key:
          type: string
          nullable: true
          maxLength: 1024
    AssignmentResponse:
      type: object
      required: [assigned]
      description: Response from POST /poll — either a subtask assignment or empty.
      properties:
        assigned:
          type: boolean
        subtask:
          $ref: '#/components/schemas/SubTaskRead'
          nullable: true
        assignment_token:
          type: string
          format: uuid
          nullable: true
        repo_id:
          type: string
          nullable: true
        revision:
          type: string
          nullable: true
        storage_config:
          $ref: '#/components/schemas/StorageConfig'
...